requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.9
oauthlib==3.1.0
python-twitter==3.5
requests==2.24.0
requests-oauthlib==1.3.0
soupsieve==2.0.1
urllib3==1.25.9
